Update patch set 1
Patch Set 1:

I have seen cases where the same input is used for several instructions, some of which need a register, and some a constant/register.  In this case, the register allocator reports that the location is in the register, even if the actual input is constant.  This can lead to generating code that could use constants rather than registers.

Patch-set: 1