Update patch set 1
Patch Set 1:

> I have seen cases where the same input is used for several
 > instructions, some of which need a register, and some a
 > constant/register.  In this case, the register allocator reports
 > that the location is in the register, even if the actual input is
 > constant.  This can lead to generating code that could use
 > constants rather than registers.

If you find such a case with the current implementation of Location::RegisterOrConstant(), i.e requesting constant locations for constant inputs, please report a bug against the RA.

However, to me the name RegisterOrConstant() implies that the code generator doesn't care. If it does, it should not be calling that function.

Patch-set: 1