Update patch set 1
Patch Set 1:

Geoffray, with custom binaries there are many ways to get the logs without loss. But we need to test release builds. And the most tough are the releases that are limited to be based on the AOSP sources only. That is why we badly need a solution merged in AOSP.

Patch-set: 1