Update patch set 2
Patch Set 2:

> > Do we also want to perform this split on Intel? and for MIPS? If
 > > not I can move that to the arm specific optimisations.
 > 
 > Only MIPSR6 (but not MIPSR2-) can benefit from it as it has the
 > [D]LSA instruction that adds a register to another register
 > multiplied by 2/4/8/16.

I mean, benefit as in using a single instruction ([D]SLA) for this. 

As far as I understand, shift + add/sub shouldn't be worse than load constant + mul. I'd expect same or better (loading constants is more expensive on MIPS64 (2 instructions needed to load a large power of two into a register) and multiplication may not be done in one clock (I know some MIPS cores do multiplication in more clocks depending on the multiplicands)).

So, I'd say this should be OK to do for MIPS.

Patch-set: 2