Update patch set 1
Patch Set 1:

> This bug affects A53's.  ART currently assumes all arm64 are A53's;
 > while kryo is custom arm64 design - not A53.
 > 
 > This was the simplest way to bypass the A53 checks for kryo.

I fully understand that - I can read reviews. :-) I just want this to be documented here (it was a mistake not doing that for Denver).

 > https://www.qualcomm.com/products/snapdragon/processors/comparison

Is that the best we can do? If so, please add a comment like "Kryo is not based on the ARM Cortex A53 and thus does not need erratum handling. [link]"

Patch-set: 1
Label: Code-Review=0