Update patch set 6
Patch Set 6:

> I understand the problem; I am not sure how to fix it

Well, if we are okay with this as public API (but I am not sure we should be), boolean_simplifier works as is since it used that method before: only your instruction_simplifier need to be deal with a returned instruction may already be in the HIR (see some of the surrounding code).

We could also make the general API a bit more elaborate, and pass in an instruction where a possibly new instruction should be inserted. This seems cleaner, but requires rewriting both the client sites.

I realize this may have become a bit more elaborate than you signed up for :-), but it is hopefully for the common good.

Patch-set: 6