Update patch set 2
Patch Set 2:

I believe this should basically work, but ...

Is there an Intel recommendation to this effect someplace?  It's at odds with http://www.cl.cam.ac.uk/~pes20/cpp/cpp0xmappings.html, which doesn't even list that alternative.  There are probably good reasons for that (see below) but it's still weird, and makes me uncomfortable.

The worst-case scenario is clearly that this changes with every new microachitecture.  I'm pretty sure this trade-off was different when I measured it last, about a decade ago.

The other concern I have is that the new version is, I believe, a very unexpected write to 0(sp).  We now need to convince ourselves that that cache line is never shared with another thread, and thus can't introduce coherence misses.  And that it doesn't dirty enough extra cache lines to matter.  That's probably true for Java, but it's yet another thing to keep in mind for the next clever optimization that might invalidate it.

Presumably we are generating these fences on x86 primarily for volatile stores.  Can we instead just compile those to an xchg instruction for the same performance benefit, without the hazards?  I believe that has always been listed as a standard x86 translation of a volatile store.

Patch-set: 2
Label: Verified=0