Update patch set 3
Patch Set 3:

This is indeed guaranteed not to be shared with another thread in, at least naively compiled, Java code.  (Except maybe if the other thread is the garbage collector, which is probably not a big deal.)

But thread stacks are routinely accessed by multiple threads in C++ code.  Just invoke a parallel sort routine on a stack allocated array.  Can we now generate such a scenario in Java with JNI?  I don't know.  Are we ever going to add optimizations that will allow pure Java code to do this?  I haven't a clue ...

This feels to me like we're planting a land mine for future implementers.  We might not ever detonate it, but my personal preference would be not to plant it if we can avoid it.  Any kind of invented store just makes me really nervous in a modern multicore context.  The xchg option seems MUCH cleaner, though clearly also more work.  But I think it's probably work that needs to be done for other architectures anyway.

Do we know the performance difference here is portable across different implementations of the architecture?

IIRC, this is actually an ancient SPARC hack that predates their fence instructions.  That doesn't make me a fan.

Patch-set: 3