Update patch set 1
Patch Set 1:

(5 comments)

> Any way to avoid the copying of the ArtMethod instead? The fix
 > looks very brittle, and the mistake of doing pointer comparison
 > easy to repeat.

@ngeoffray, @shertz

We need to copy the methods because they must have a different method index then their original interface versions. This is because in a class the method_index is used as the vtable position but in interfaces it is used as the index in the iftable. Since these are different and the compiler depends upon this to create invoke-virtuals and invoke-supers we must copy.

Furthermore I have made sure to hide all the pointer comparisons behind the IsInMethod function of breakpoint. The only other places the GetCanonicalMethod function is used is when communicating to the client which needs them to be the same for display purposes.

Patch-set: 1
Label: Verified=0