Update patch set 4
Patch Set 4:

I was thinking about the semantics of XCHG last night. An XCHG is a read of the value, followed by a write.  Is that legal for Java?  I believe that in C, you aren't allowed to read a volatile unless directed by the program.  Is that also a consideration for Java?  It might not be a consideration, as there are no 'pointers to device memory' in Java.

Comments by a Java expert?

Patch-set: 4