Update patch set 4
Patch Set 4:

I would still like to at least see a statement of the form: This improved application/benchmark X by N percent, where X is something more real than running the instruction in a loop.  That's particularly important here, I think since, at least in my experience, the timing of these instructions often depends a lot on adjacent memory references.

My previous reply, which inadvertently didn't make it here:

Thanks!  I don't agree with all of Paul's comments, but I've become convinced it's OK to proceed with this CL if we can get a bit more evidence that this makes a difference on real code.

Some replies:

Mark's question:

I believe it's entirely fine to add the extra load for XCHG.  Since you already need exclusive cache line access, I don't think that's an issue.  Java volatiles correspond to C/C++ atomics, and have little to do with C/C++ vollatile.

Paul's comments:

I agree that ARM constraints shouldn't influence x86 code generation.  I only cited the code as an example for how to do this.  I also agree we're not talking about C and C++ volatiles, which are an entirely different beast.  C/C++ atomics are much more relevant.

The JSR 133 cookbook was a great step forward at the time, but it doesn't fully address all of the subtleties of non-multi-copy-atomic-store architectures.  As a result, I don't think it's a great model anymore, in spite of the fact that we're not dealing with any of those more difficult architectures here.  http://www.cl.cam.ac.uk/~pes20/cpp/cpp0xmappings.html is now a better guideline, in spite of the relative lack of intuition, if you remember that Java volatile = C++ seq_cst operation.  However, for x86, I think either works.  (PowerPC and architectures with acquire/release instructions cause issues with the Cookbook model, which are hinted at by some recently added text.  The Cookbook model also doesn't address the issue of multiple incompatible translations, which I think is a serious problem requiring ABI attention.)

Separating the fence and the store in the IR works if you plan to generate "all-or-nothing" fences that order all preceding accesses with respect to all later ones.  It's not a natural thing to do if the generated instruction orders, say, a single store with respect to all earlier accesses.  You can't really separate the two in any case.  The latter approach, at least in theory, constrains the hardware a lot less.  And we basically have to deal with both.  (HotSpot does to, I believe.  But my impression is that it deals much better with the former.)

I also agree with the comment about few registers, though I wonder how often the value used in a volatile store remains live.  For the normal use cases I can think of, that doesn't seem to be the case.

Optimization with an adjacent fence instruction seems to be a very special case issue.  In general, on x86, I only need to guarantee that a volatile store is followed by a fence before the next volatile load.  This means that if, for example, I have a volatile store, followed by non-volatile operations, followed by either another volatile store or an atomic RMW, I can drop the fence for the first store.  In general, this analysis doesn't seem to have a lot to do with the instruction used for the volatile stores that do need to enforce ordering.

The kind of cache conflicts I was worried about are:

1) At some future point, we decide to stack allocate a Java object that is shared with another thread, but known not to outlive the current frame.

2) Some C code starts another thread with a pointer to the local frame, and then invokes Java code while that thread is still running.  The Java stack isn't far enough away from 

But I agree that in both cases this is no worse than Java code unexpectedly pushing and popping stuff on its local stack, so we can probably live with it.

Patch-set: 4
Label: Verified=0