Update patch set 1
Patch Set 1:

This CL addresses only the x86 case – because it was written to exercise the bug fixed in https://android-review.googlesource.com/#/c/184784 – but I intend to extend it to other architectures.

So far, I have been using the fact that art:Location::kOutputOverlap is the default value in art::LocationSummary::SetOut, but the more I think about it, the more I'm convinced we should drop this default value in art::LocationSummary::SetOut.  What do you think?

Patch-set: 1