Update patch set 3
Patch Set 3: -Verified

I'm saying that the simplistic command to run them (which works on other benchmarks) doesn't work because it can't find the shared libraries. I don't know whether that is because:
1) the command is wrong.
2) I have broken Vogar.
3) I have to do something else, e.g. build a specific target, before running Vogar.

The main reason I can't determine that is because there aren't any instructions on how the benchmarks should be run. If there were then I would have verified that those instructions still worked but without them I would be shooting in the dark.

If you can let me know how to run the benchmarks then I will do that to verify that they still work.

Patch-set: 3