Update patch set 1
Patch Set 1:

> I don't really see the claimed simplification.

It's mostly makes it simpler for the invoke-super cl. Basically there are several places where we needed to filter out a lot of different types of methods (i.e. miranda, default, etc) and there were so many that it was getting confusing. 

This CL means there are far fewer places where we need to filter out stuff. It also removes the need for repetition in several places.

As I said it also makes the class object slightly smaller too.

Patch-set: 1
Label: Verified=0