Update patch set 1
Patch Set 1: Code-Review-1

It is *not* enough to just re-verify. If a dependency was *used* (class layout etc), then all your quickened or compiled code is potentially *wrong*, but it will still verify with the new layout.

Your approach is overly complex for something that we usually cannot recover from. The suggested approach was to throw away the oat files when *anything* changes and let the interpreter or JIT take care of it.

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1