Update patch set 1
Patch Set 1:

> (1 comment)

Dex constraint A14 (or the part of it that we verify). Moving the check early returned an ICCE error instead of the hard failure in case invoke-interface calls <init> or <clinit>. Testing the fix now, which is to instead postponing the push of the resolved method into the dex cache.

Patch-set: 1