Update patch set 1
Patch Set 1: Code-Review-2

I rather this doesn't get in (in this form).

The commit message is ambiguous. It's your company that has a "native bridge." You probably mean "native transition."

This isn't a clear win except for fast-native (otherwise I'd blame GCC or Clang for crappy code). Where are your benchmark results for common apps that show the ratio of native to fast-native calls?

Please stop hacking architecture-specific code into files that are *clearly* designed to be architecture-dependent. Change the general code and extend the assembler/calling convention code, or I will go and -2 any such change.

When you do, also consider an infrastructure for profiled results and invalidation. While both the synchronized and the fast-native state can change, that shouldn't happen often. So if you integrate this into JIT profiling, you can remove the cmp-and-branch (and fall back to Generic JNI in the bad case, when loading the class).

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2