Update patch set 24
Patch Set 24:

(11 comments)

As discussed offline (with Alex) it's unlikely to be worth it taking a 76k/app regression in memory, especially since none of the current apps are using the new invoke-super functionality.

So perhaps for now, it can look up the supers recursively (which will be slower) and if we want to speed it up later then one right way which doesn't involve a memory regression would be to add compiler support.

Patch-set: 24
Label: Verified=0