Update patch set 1
Patch Set 1:

> I personally don't believe log levels make that much difference in
 > diagnosing errors.

Interesting. Just curious, what would you personally use it for?

If you believe log levels make a significant
 > difference in diagnosing errors, then we should not turn what
 > should be warning messages into info messages.

The point is that with the move to JIT it being a warning is harmful as it's expected behavior for some situations.

 Even though the
 > particular error message you are seeing is misleading until such
 > time as the cleaner fix is implemented, there are many other cases
 > when MakeUpToDate can fail, and those messages should be reported
 > as warning messages, not info messages.
 > 
 > Is it so hard to fix the underlying issue?
 > 
 > If you want an easy hack to bother people less while you work on
 > resolving the underlying issue, I would prefer you remove the
 > entire call to 'MakeUpToDate'.

Maybe it's the larger picture I failed to communicate? MakeUpToDate on itself is fine, and will be called for application loaded dex files, for which the application has the rights to compile them in its directory. However, it's the application's dex files themselves for which it does not have the rights.

The proper fix is to be able to differentiate between the two. Not hard, just more involved then a %s/WARNING/INFO.

Patch-set: 1