Update patch set 1
Patch Set 1:

> If sudden reset or reboot by user happens during compilation, the
 > oat file is not created normally. In this case, since some part of
 > the file is not flushed, the file is properly not working at
 > runtime.

Do we detect the problem with the oat file when we try to open it at runtime, or does it lead to a crash or bad behavior? Do you have examples of log messages from ART around the time when the problem using the oat file occurs?

Patch-set: 1
Reviewer: Richard Uhler <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0