Update patch set 1
Patch Set 1:

Yeah, it would be cool ninja can show clearer log. I think if we really want the equivalent to $? used here, we could file a feature request to ninja, so it benefits all other targets as well.

Andreas seems to be OK with this so I'd like to submit this CL as is. I'll wait until tomorrow just in case someone have objections.

Patch-set: 1