Update patch set 2
Patch Set 2:

The bitmap memmap is page aligned but the bitmap_size_ field is the number of words in the bitmap. So we will have most bytes per bit * bits per word of stuff past the end of the image objects, however since we round up the image size to page size, the bitmap does not have any bits that could cover other spaces or objects.

Patch-set: 2