Update patch set 2
Patch Set 2:

> > Uploaded patch set 2.
 > 
 > For the case shown in the test, since the do-while loop is always
 > entered, is it more straightforward to find it's always entered and
 > don't generate the taken-test blocks at all? Essentially the loop
 > header is the loop body in this case.

That is an excellent observation. If we would test that there is nothing else in a loop body proper, we could omit that test indeed. However, without adding more complex and error prone tests, we cannot assume that.

It is a bit hard to express a Java example that has *and* stuff in the header *and* stuff in the body, but it does not seem impossible. In such cases, we need to generate the top test.

It is a lot easier to just do the right thing, and have a cleanup phase remove the blocks that are redundant afterwards.

Patch-set: 2