Update patch set 2
Patch Set 2:

Also about the dynamic bce code, is it possible for TransformLoopHeaderForBCE() to possibly fail? i.e. the condition test to see if the loop is entered depends on some definitions in the loop header, so the test can't be done in an added if_block in front of the loop header. Maybe we have some restrictions so that won't happen, just want to make sure. Adding some DCHECK's will also help.

Patch-set: 2
Label: Verified=0