Update patch set 1
Patch Set 1:

If we consider to take this: the change is at the wrong point.

The PMS doesn't know the size of the code, so it can't decide.

But dex2oat already does a size decision for swap. Why are you polluting CompilerOptions with this, and have a second own collecting of size stats?

Patch-set: 1
Label: Verified=0