Update patch set 1
Patch Set 1:

Is this idea going to work?
I assume the method array is pretty much the equivalent of the vtable, and thus good candidate to use as unique identifier for the class.
The most important question is whether the GC can move it.  It seems to me the answer might fortunately be no.

Patch-set: 1