Update patch set 2
Patch Set 2:

> I'm seriously missing something here.... builder.cc should be able
 > to resolve the invoke and make it a HInvokeStaticOrDirect. If it
 > cannot, HInvokeUnresolved will be used instead.
 > 
 > That's basically the way I understand it. Clearly I'm missing
 > something, othwerwise we wouldn't need so much code for for those
 > simple two cases.

I put the code to devirtualize the HInvokeSuperInterface to HInvokeStaticOrDirect in sharpening.cc based on the name and some of the comments in the file. It wasn't until near the end of creating this that I realized that the corresponding transformation for regular invoke-super occurs during the creating of the graph in builder.cc.

By the way, why does this devirtualization occur there? It seems to me that this sort of transformation should not be occuring as we are generating the graph. Furthermore the comments in sharpening.cc seem to support me in this (why I put it in there in the first place).

Patch-set: 2