Update patch set 1
Patch Set 1: Code-Review-1

This isn't quite right for x86.  You need to check for SSE 4.2 and the popcnt bit to see if the popcnt instruction is supported.  If so, this can be used in 32 bit mode too (two popcntl's + add)

This needs to add a new feature and have the feature set for the variants.  Silvermont supports popcnt

I am not sure if there are any x86 variants with SSE4.2 support that don't support popcnt.  I will ask around.

Patch-set: 1
Label: Code-Review=-1