Update patch set 3
Patch Set 3:

Overall, looks good.  There is one thing that caught my attention, but after giving it considerable thought I can't come up with a good reason to ask you to change it.  It relates to the use of REFRESH_IBASE.  Once we have an alt version of mterp active, the REFRESH_IBASE operation is the point at which you can switch between the two.  In the Dalvik days (and with the mterp ARM target), IBASE refreshes are limited to potential safe points.  Thus, we can only transition at safe points.  For x86, though, you are using REFRESH_IBASE in two ways - to transition at a safe point, but also to essentially unspill the IBASE register after a call to a runtime support routine.  However, the version of IBASE that is saved in thread local storage can change during the time the support routine is running, so you may refresh to a different value than the one you started with.

This means that the x86 mterp will have a different set of possible transition points than ARM mterp will.

But, I don't believe this hurts anything.  We could have an UNSPILL_IBASE macro which truly unspills rather than refreshes for x86 to distinguish the two usages, but again - I can't see how this would cause a problem.

So, let's keep what you've got - but also keep this slight behavioral difference in mind.  We may need to revisit it later.

Patch-set: 3