Update patch set 1
Patch Set 1:

> How exactly do you want to test this? Run the 004-UnsafeTest in the
 > "gcstress" configuration?

Something like that, I guess. It doesn't have to be 100% rate, but setting up something that triggers a GC so that the marking becomes relevant would be nice.

(I'm fine merging this, as the test is probably a little bit involved, but I think the necessity of this CL shows that our testing isn't good enough, yet)

Patch-set: 1