Update patch set 1
Patch Set 1:

> I think the root cause of the problem is that we are using the
 > --generate-debug-info flag for 2 different functionality. We use it
 > for generate dwarf debug info and pass it to a native debugger
 > (LLDB) through the JIT debug interface what should work both on
 > user and userdebug builds. On the other hand we also use the flag
 > to generate the perf map file what is only working on userdebug
 > builds and crashing on user builds.
 > 
 > I think we have a few options:
 > 1. Change the pref map file generation to work on user builds with
 > either making the failures a soft failure or with moving the perf
 > file to some more accessible location (e.g. application directory)
 > 2. Use a different flag to enable the perf map file generation and
 > the debug info generation so they can be enabled separately. That
 > way we can say that perf is only working on userdebug builds.
 > 3. Restrict the perf map generation to userdebug (or debug) build
 > only.
 > 
 > Which solution would you prefer?

I would generate/save the perf map under /data/misc/trace/ only in userdebug/eng build. That's a directory where the apps save their traces and works only in userdebug/eng builds. So you don't need to bother with disabling selinux. 

With respect to using the app directory for the debug info. You should check with the fraemwork and/or security team as well. I'd prefer to have another VMRuntime method which will tell the runtime the file where this info should be saved. ActivityThread can pass this file only if the app is debuggable. This way, you don't depend on other flags being enabled or not and the intention is much clearer. (Also, the runtime should be free of making decisions where to put files...)

Patch-set: 1