Update patch set 4
Patch Set 4:

> Aart, what's the diff between the original and the fix?

I uploaded the original as PS1 and the modified as PS2. Basically I no longer rely on whether that 2nd array ref is inlined or not (not something I wanted to test, but in the method above it is guaranteed to happen for all other types, so I had to single this one out anyway).

Patch-set: 4