Update patch set 1
Patch Set 1:

For my purpose (native debugging) we don't have to save the debug info to disk because we just pass it to the debugger (in memory) and then the debugger handles it.

The main question is where should I put the perf map file in case of user builds? It can be specified either by the runtime or by the activity thread (I don't really care) but I don't see any location where I can store it.

Optionally what is the general opinion about saving the perf map file only in case of userdebug/eng builds? (Can I check it in the runtime?)

Patch-set: 1