Update patch set 5
Patch Set 5:

I looked at the code a bit.  I am worried about the path from native code to interpreter (EnterInterpreterFromEntryPoint).  This does not go through ArtMethod::Invoke so we would have to handle it separately as well.

It seems difficult to avoid creation of the shadow frame.  EnterInterpreterFromEntryPoint already has it and ArtInterpreterToCompiledCodeBridge needs it anyway.  And I would like to keep ArtInterpreterToCompiledCodeBridge since it handles static initialization of classes if needed.  In any case, all of this happens only on the first invocation, the other invocations will be direct to the compiled code, so I am not too worried about the overhead.

So overall I am tempted to keep the Siva's version as is if Nicolas and Andreas are happy with.

Patch-set: 5
Label: Verified=0