Update patch set 1
Patch Set 1:

I am happy with putting the perf map file under /data/misc/trace/$(get_process_name()).perf but it is still only accessible on userdebug builds.

If there is no way to detect if we are running on a user or a userdebug build then I think we have to make it a soft failure so it won't crash on user builds. I think it would mean that we can run perf even on user builds but we won't have data for jitted code (the UI for perf can emit a warning for it saying JIT perf requires userdebug build). I think it is better then saying that perf is not supported on user builds at all.

Patch-set: 1
Label: Verified=0