Update patch set 4
Patch Set 4:

> > I wish, but I don't know how to do it reliably. When I had it
 > partially tied down, it was pretty flaky. I can compile two sets of
 > core files and compare, but it will not reliable trigger on
 > regression.
 > 
 > Would it be not-perfectly-reliable but somewhat helpful than
 > nothing (eg to find some run-to-run variations)?

Of course it would be better than nothing. However, even the setup is already complicated, as our tests don't know how to compile a core image. I am not sure at this point whether a gtest is the right option, or some explicit Make rule...

I'd prefer to put this into a follow-up at this point.

Patch-set: 4