Update patch set 3
Patch Set 3:

> > Do we need to add a flag to say there is no compiled code?
 > 
 > That's the VerifyAtRuntime part from the header.
 > 
 > > Does the
 > > oat file not already have that information?
 > 
 > We do not have this flag, as it was never useful. Traditionally we
 > haven't done derived this info (see PIC, which could check some
 > things). I agree though that there should at least be DCHECKs to
 > confirm that the flag agrees with the oat content.

I think my question wasn't clear. This CL adds a flag to the oat file to tell the oat file assistant not to check the image checksum. But isn't there already information in the oat file saying which methods are compiled? Couldn't the oat file assistant already use that information to know if it can skip the image checksum check? Then you could skip the image checksum in other cases, such as if you decided not to compile methods. And then you don't need to add a flag that could be set incorrectly.

I don't know the oat file format well enough to know what parts of it depend on the image, and how much it costs to figure that out from the existing oat file, which is why I ask.

Patch-set: 3
Label: Verified=0